requirements.txt
playwright
requests
beautifulsoup4
geopandas
python-telegram-bot==13.15
python-dotenv
